26.12.2019

+++ VehicleDealershipDB/veh/Stored Procedures/sp_insert_vehicle.sql
@@ -14,7 +14,7 @@ CREATE PROCEDURE [veh].[sp_insert_vehicle]
 	@is_new bit,
 	@location int,
 	@engine_no nvarchar(20),
-	@engine_cc float,
+	@engine_cc int,
 	@mileage int,
 	@vehicle_sale int,
 	@consignment_mortgate bit,
